YoutubeTranscriber/YoutubeToPPT/requirements.txt
google-generativeai
youtube-transcript-api
streamlit